Wait for network config before activating libvirt socket
Wait for the network configuration to complete before trying to activate
the libvirt socket. This prevents systemd binding to the specified IP,
then having the interface either come up or reconfigure.

Due to the race-y nature of the network config (at least with
NetworkManager) and the socket activation this would mean <50% of the
machines rebooting would be affected. After this point the socket is
"up" from systemd's POV however anything (including telnet) trying to
open the port will find it closed until an administrator restarts the
socket unit